% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moi.R
\name{applyMOI}
\alias{applyMOI}
\title{Applying modes of inheritance}
\usage{
applyMOI(X, modes)
}
\arguments{
\item{X}{Matrix or data frame of SNPs coded as 0,1,2}

\item{modes}{Character vector of MOIs that can be obtained using
\code{\link{MOI}}.}
}
\value{
A matrix or data frame of SNPs using the given MOIs
}
\description{
Transform SNPs according to the supplied modes of inheritance (MOIs).
}
