% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calcNMSE}
\alias{calcNMSE}
\title{Calculate the NMSE}
\usage{
calcNMSE(preds, y)
}
\arguments{
\item{preds}{Numeric vector of predictions}

\item{y}{True outcomes}
}
\value{
The NMSE
}
\description{
Computation of the normalized mean squared error.
}
