% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bits.R
\name{calcNoPossibleTerms}
\alias{calcNoPossibleTerms}
\title{Number of possible terms}
\usage{
calcNoPossibleTerms(p, max.vars, negate = TRUE)
}
\arguments{
\item{p}{Number of input variables}

\item{max.vars}{Maximum interaction order}

\item{negate}{Shall negations of input variables also be considered?}
}
\value{
Number of possible terms
}
\description{
Calculates the possible number of marginal terms and interaction terms.
}
