% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bits.R
\name{cv.BITS}
\alias{cv.BITS}
\title{Cross-validation for BITS}
\usage{
cv.BITS(
  X,
  y,
  nfolds = 5,
  parallel = TRUE,
  gmin = function(gmax) 0.001 * gmax,
  gsteps = 50,
  choose = "min",
  metric = "auc",
  ...
)
}
\arguments{
\item{X}{Matrix or data frame of \eqn{p} input variables}

\item{y}{Numeric vector of a binary or continuous outcome}

\item{nfolds}{Number of cross-validation folds}

\item{parallel}{Shall parallel computation be enabled?}

\item{gmin}{Ratio of the minimum \code{gamma} value/maximum \code{gamma} value}

\item{gsteps}{Number of \code{gamma} values}

\item{choose}{\code{"min"} for choosing the model yielding the minimum
cross-validation error (default). Alternatively, \code{"1se"} can be used
to obtain the least complex model (high \code{gamma} and high \code{lambda})
such that its cross-validation error is within 1 standard error of the
minimum cross-validation error.}

\item{metric}{For binary outcomes, either "auc" (default) to use the area
under the curve as performance metric or "dev" to use the deviance as
performance metric}

\item{...}{Arguments passed to \code{\link{BITS.complete}}}
}
\value{
A \code{BITS} model. Additionally, \code{$cv} contains a list
  consisting of
 \item{\code{best.gamma}}{The identified optimal value for \code{gamma}}
 \item{\code{best.s}}{The identified optimal value for \code{lambda}}
 \item{\code{cv.res}}{A data frame with cross-validation errors for
   different values of \code{gamma} and \code{lambda}}
}
\description{
Performing k-fold cross-validation to obtain a BITS model with tuned
\code{gamma} and \code{lambda} penalty parameters.
}
\details{
For parallel computing, a cluster has to be registered beforehand, e.g., via
\preformatted{
 cl <- parallel::makeCluster(4) # Using 4 threads
 doParallel::registerDoParallel(cl)
}
}
