% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bits.R
\name{gammaPath}
\alias{gammaPath}
\title{\code{gamma}-path computation}
\usage{
gammaPath(X, y, gmin = function(gmax) 0.001 * gmax, steps = 50)
}
\arguments{
\item{X}{Matrix or data frame of \eqn{p} input variables}

\item{y}{Numeric vector of a binary or continuous outcome}

\item{gmin}{Ratio of the minimum \code{gamma} value/maximum \code{gamma} value}

\item{steps}{Number of \code{gamma} values}
}
\value{
A numeric vector containing the (decreasingly sorted) \code{gamma}
  values
}
\description{
Computes a path of \code{gamma} values to be used in \code{\link{BITS}}.
}
\details{
The maximum \code{gamma} value is computed by
\deqn{\code{gamma}_\mathrm{\max} = S_\mathrm{corr}(m_1),}
where \eqn{S_\mathrm{corr}} is the correlation score and \eqn{m_1} is the
optimal interaction tree stump containing only one input variable.

The \code{gamma}-path is then obtained by computing \code{steps} equidistant
steps on the logarithmic scale between \eqn{\code{gamma}_\mathrm{\max}} and
\eqn{\code{gmin} \cdot \code{gamma}_\mathrm{\max}}.
}
\references{
\itemize{
  \item Lau, M., Schikowski, T. & Schwender, H. (2023).
  Boosting interaction tree stumps. To be submitted.
}
}
