% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bits.R
\name{get.ideal.model}
\alias{get.ideal.model}
\title{Ideal hyperparameters}
\usage{
get.ideal.model(models, X, y, choose = "min", metric = "auc")
}
\arguments{
\item{models}{An object of class \code{BITS.list}, which is a list of fitted
\code{BITS} models. This object can be obtained using
\code{\link{BITS.complete}}.}

\item{X}{Matrix or data frame of \eqn{p} input variables (of validation data)}

\item{y}{Numeric vector of a binary or continuous outcome (of validation data)}

\item{choose}{\code{"min"} for choosing the \code{gamma} and \code{lambda}
values yielding the minimum cross-validation error (default). Alternatively,
\code{"1se"} can be used to obtain the least complex model
(high \code{gamma} and high \code{lambda}) such that its cross-validation
error is within 1 standard error of the minimum cross-validation error.}

\item{metric}{For binary outcomes, either "auc" (default) to use the area
under the curve as performance metric or "dev" to use the deviance as
performance metric}
}
\value{
A list containing
  \item{\code{val.res}}{A data frame with validation errors for
    different values of \code{gamma} and \code{lambda}}
  \item{\code{best.g}}{The identified optimal value for \code{gamma}}
  \item{\code{best.s}}{The identified optimal value for \code{lambda}}
}
\description{
Obtain the optimal \code{gamma} and \code{lambda} penalties.
Independent validation data should be supplied to unbiasedly identify the
right values.
}
