% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bits.R
\name{get.ideal.penalty}
\alias{get.ideal.penalty}
\title{Ideal term inclusion penalty}
\usage{
get.ideal.penalty(model, X, y, choose = "min", metric = "auc")
}
\arguments{
\item{model}{A fitted \code{BITS} model}

\item{X}{Matrix or data frame of \eqn{p} input variables (of validation data)}

\item{y}{Numeric vector of a binary or continuous outcome (of validation data)}

\item{choose}{\code{"min"} for choosing the \code{lambda} value yielding the
minimum cross-validation error (default). Alternatively, \code{"1se"} can
be used to obtain the highest \code{lambda} value such that its
cross-validation error is within 1 standard error of the minimum
cross-validation error.}

\item{metric}{For binary outcomes, either "auc" (default) to use the area
under the curve as performance metric or "dev" to use the deviance as
performance metric}
}
\value{
A list containing
  \item{\code{val.res}}{A data frame with validation errors for
   different values of \code{lambda}}
  \item{\code{best.s}}{The identified optimal value for \code{lambda}}
}
\description{
Obtain the optimal \code{lambda} penalty for including identified terms in
the final relaxed lasso/elastic net model.
Independent validation data should be supplied to unbiasedly identify the
right value.
}
