% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bits.R
\name{get.included.vars}
\alias{get.included.vars}
\title{Extract terms from a \code{BITS} model}
\usage{
get.included.vars(model)
}
\arguments{
\item{model}{A fitted \code{BITS} model}
}
\value{
A matrix of the included terms. Each row corresponds
  to a single term and each entry corresponds to the column index in \code{X}.
  Negative values indicate negations. Missing values mean that the term
  does not contain any more variables.
}
\description{
If the term inclusion/lasso penalty \code{lambda} has been set in
\code{model$s}, the included terms can be extracted using this function.
}
