% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moi.R
\name{MOI}
\alias{MOI}
\title{Modes of inheritance}
\usage{
MOI(X, y, fast = FALSE)
}
\arguments{
\item{X}{Matrix or data frame of \eqn{p} SNPs coded as 0,1,2}

\item{y}{Numeric vector of a binary or continuous outcome}

\item{fast}{Shall fastglm instead of glm be employed for model fitting?
Default is \code{FALSE}. However, for high-dimensional problems, setting
\code{fast = TRUE} is recommended.}
}
\value{
A character vector of length \eqn{p} containing the identified MOI,
  where \code{"A"} is the additive MOI, \code{"D"} is the dominant MOI, and
  \code{"R"} is the recessive MOI
}
\description{
Determining the most plausible modes of inheritance (MOIs).
}
\details{
For each considered SNP (single nucleotide polymorphism) and each considered
MOI (additive, dominant, and recessive), likelihood-ratio tests are carried
out testing the association of the SNP in the considered MOI with the
outcome. For each SNP, the MOI that yields the highest association with the
outcome, i.e., the MOI that yields the lowest p-value, is returned.
}
\references{
\itemize{
  \item Lau, M., Schikowski, T. & Schwender, H. (2023).
  Boosting interaction tree stumps. To be submitted.
  \item Scherer, N., Sekula, P., Pfaffelhuber, P. & Schlosser, P. (2021).
  pgainsim: an R-package to assess the mode of inheritance for quantitative
  trait loci in GWAS. Bioinformatics, 37(18):3061–3063.
  \doi{https://doi.org/10.1093/bioinformatics/btab150}
  \item Petersen, A. K., Krumsiek, J., Wägele, B., Theis, F. J., Wichmann,
  H.-E., Gieger, C. & Suhre, K. (2012). On the hypothesis-free testing of
  metabolite ratios in genome-wide and metabolome-wide association studies.
  BMC Bioinformatics, 13:120, 2012.
  \doi{https://doi.org/10.1186/1471-2105-13-120}
}
}
