% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bits.R
\name{plot.val.performance}
\alias{plot.val.performance}
\title{Hyperparameter plot}
\usage{
\method{plot}{val.performance}(x, ...)
}
\arguments{
\item{x}{An object of class \code{val.performance}, which is a data
frame with validation data performances for different \code{gamma} and
\code{lambda} values. Can be obtained via \code{\link{BITS.complete}}.}

\item{...}{Additional unused arguments}
}
\value{
A \code{ggplot} depicting the validation data performances
}
\description{
Creates a plot showing the validation data performances for different
hyperparameter choices.
}
