% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bits.R
\name{predict.BITS}
\alias{predict.BITS}
\alias{predict.intstump}
\title{Prediction for BITS models}
\usage{
\method{predict}{BITS}(object, X, ...)

\method{predict}{intstump}(object, X, ...)
}
\arguments{
\item{object}{Fitted BITS model. Usually a product of a call
to \code{\link{BITS}} or \code{\link{cv.BITS}}.}

\item{X}{Matrix or data frame of input data. This
object should correspond to the input data for fitting
the model.}

\item{...}{Additional unused arguments}
}
\value{
A numeric vector of predictions. For binary outcomes,
  this is a vector with estimates for \eqn{P(Y=1 \mid X = x)}.
}
\description{
Supply new input data for predicting the outcome with a fitted
BITS model.
}
