% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bits.R
\name{print.BITS}
\alias{print.BITS}
\title{Print BITS model}
\usage{
\method{print}{BITS}(
  x,
  X.names = NULL,
  MOI = NULL,
  n.digits = 2,
  negation = "^c",
  SNP = FALSE,
  latex = FALSE,
  ...
)
}
\arguments{
\item{x}{Fitted BITS model. Usually a product of a call
to \code{\link{BITS}} or \code{\link{cv.BITS}}.}

\item{X.names}{Optional character vector of original predictor names.
E.g., \code{X.names=colnames(X)}, where \code{X} is the original dataframe
or matrix of predictor values.}

\item{MOI}{Optional character vector of modes of inheritance. Usually a
product of a call to \code{\link{MOI}}.}

\item{n.digits}{Number of digits to print in the regression coefficients}

\item{negation}{How to indicate variable negations? \code{negation="^c"}
simply indicates negations using superscript c and \code{negation="full"}
prints negations in more detail with the negation offset.}

\item{SNP}{Boolean value to indicate if input variables should be printed as
SNP or more generically as X}

\item{latex}{Boolean value to indicate if model should be printed in LaTeX
format}

\item{...}{Additional unused arguments}
}
\description{
Prints the fitted BITS model as a linear interaction model
}
